refactor: relocate frontend-only utils to shared module
