+++ frontend/src/components/pet/SendPetModal.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { usePetsContract } from '@shared/core';
 import { petsContractParams } from '../../petsContractParams';
 import TransactionStatus from '../ui/TransactionStatus';
-import { isValidEthAddress } from '../../utils/isValidEthAddress';
+import { isValidEthAddress } from '@shared/core';
 import './SendPetModal.css';
 
 interface SendPetModalProps {
