+++ frontend/src/components/pet/interactions/BreedPanel.tsx
@@ -7,10 +7,10 @@ import {
     usePetsContract,
     useWatchPetsContract,
     type BreedSuccessPayload,
+    getReadyPets,
 } from '@shared/core';
 import { petsContractParams } from '../../../petsContractParams';
 import { DASHBOARD_HOME } from '../../../constants/interactionRoutes';
-import { getReadyPets } from '../../../utils/readyPets';
 import { useWriteContractErrorState } from '../../../hooks/useWriteContractErrorState';
 
 export type BreedPanelProps = {
