+++ shared/src/utils/ethereum/isValidEthAddress.ts
