+++ shared/src/utils/ethereum/readyPets.ts
@@ -1,4 +1,4 @@
-import type { Pet } from '@shared/core';
+import type { Pet } from '../../hooks/ethereum/usePetsContract';
 
 /** Pet id + record, only for pets that are ready for interaction. */
 export type ReadyPetOption = { id: bigint; pet: Pet };
@@ -6,7 +6,7 @@ export type ReadyPetOption = { id: bigint; pet: Pet };
 export function getReadyPets(
     petIds: bigint[],
     pets: Pet[],
-    isReady: (readyTime: bigint) => boolean,
+    isReady: (readyTime: bigint) => boolean
 ): ReadyPetOption[] {
     return petIds
         .map((id, index) => ({ id, pet: pets[index] }))
