+++ shared/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './ethereum';
 export * from './solana';
